🚚 moved the estimation problem creation into the respective benchmark file
